+++ ng/audio.cpp
@@ -3,226 +3,353 @@
 
 #include "ng.h"
 
+// Given volume [0, 1], produce dB [-inf, 0].
+float ng::volume_to_dB (float volume) {
+	float dB;
+	if (volume > 0.0f) {
+		dB = 20.0f * std::log10(volume);
+	} else {
+		dB = -144.0f;
+		//dB = -98.0f; // 16-bit noise floor
+	}
+	return dB;
+}
+
+// Given dB [-inf, 0], produce volume [0, 1].
+float ng::dB_to_volume (float dB) {
+	float volume;
+	volume = std::pow(10.0f, dB/20.0f);
+	return volume;
+}
+
+// Convert dB to volume, mix, then convert back to dB.
+float ng::mix_dB (float dB1, float dB2) {
+	// O(time):
+	//     2 * pow(b,x) = 2 * x = 10 +/- 10
+	//     2 * div(x)   = 2 * 1 = 1
+	//         log10(x)       1 = 1
+	// Bottleneck is the dB conversion to volume.
+	// Would be 10x faster to convert to volume when clip loads, then
+	// only convert to dB when audio gets queued.
+	float volume1, volume2, volume_mix, dB_mix;
+	volume1 = dB_to_volume(dB1);
+	volume2 = dB_to_volume(dB2);
+	volume_mix = volume1 + volume2;
+	dB_mix = volume_to_dB(volume_mix);
+	return dB_mix;
+}
+
+float ng::dB_silence () {
+	return -144.0f;
+}
+
+float ng::volume_silence () {
+	return 0.0f;
+}
+
+// Load .wav file into buffer, with same spec as audio device.
 void ng::Clip::init (Audio* a, const char* file) {
 	// load audio file
-	Clip wav;
-	if (SDL_LoadWAV(file, &(wav.spec), &(wav.data), &(wav.samples)) == NULL) {
+	SDL_AudioSpec spec;
+	uint8_t* file_buffer;
+	uint32_t file_bytes;
+	if (SDL_LoadWAV(file, &spec, &file_buffer, &file_bytes) == NULL) {
 		throw std::runtime_error(SDL_GetError());
 	}
-	this->spec = a->spec;
-	this->data = NULL;
-	this->samples = 0;
 	
-	// use audio stream to convert loaded audio to desired format
+	// Use audio stream to convert loaded audio to desired format
+	this->spec = a->spec;
+	this->buffer = NULL;
+	this->samples = 0;
 	SDL_AudioStream* stream = NULL;
 	stream = SDL_NewAudioStream(
-	wav.spec.format, wav.spec.channels, wav.spec.freq,
+	spec.format, spec.channels, spec.freq,
 	this->spec.format, this->spec.channels, this->spec.freq);
 	if (stream == NULL) {
-		SDL_FreeWAV(wav.data);
+		SDL_FreeWAV(file_buffer);
 		throw std::runtime_error("unsupported audio format");
 	}
 	
-	// put audio into stream
-	if (SDL_AudioStreamPut(stream, wav.data, wav.samples) != 0) {
+	// Put audio into stream
+	if (SDL_AudioStreamPut(stream, file_buffer, static_cast<int>(file_bytes)) != 0) {
 		SDL_FreeAudioStream(stream);
-		SDL_FreeWAV(wav.data);
+		SDL_FreeWAV(file_buffer);
 		throw std::runtime_error("audio conversion failure");
 	}
 	
-	// convert audio
+	// Convert audio
 	SDL_AudioStreamFlush(stream);
 	
-	// get converted audio
-	int buffer_max = ((int)this->spec.freq * 3600) * (int)this->spec.channels;
-	uint8_t* buffer = static_cast<uint8_t*>(std::malloc(buffer_max * sizeof(uint8_t)));
-	int buffer_length = SDL_AudioStreamGet(stream, buffer, buffer_max);
+	// Get converted audio
+	// Max 10 minutes (600 seconds) of stereo 44100Hz float32 audio = 212MB!
+	int clip_max = static_cast<int>(this->spec.freq) * 600 *
+		static_cast<int>(this->spec.channels) * sizeof(float);
+	this->buffer = static_cast<float*>(std::malloc(clip_max));
+	int clip_bytes = SDL_AudioStreamGet(stream, this->buffer, clip_max);
 	SDL_FreeAudioStream(stream);
-	SDL_FreeWAV(wav.data);
-	if (buffer_length <= 0) {
-		// technically an error is -1, but we can't malloc 0 so catch that too
+	SDL_FreeWAV(file_buffer);
+	if (clip_bytes <= 0) {
+		// Technically an error is -1, but we can't realloc 0 so catch that too
 		throw std::runtime_error("audio conversion failure");
 	}
 	
-	// copy converted audio to clip
-	this->data = static_cast<uint8_t*>(std::malloc(buffer_length * sizeof(uint8_t)));
-	for (int i = 0; i < buffer_length; i++) {
-		this->data[i] = buffer[i];
-	}
-	this->samples = buffer_length;
-	std::free(buffer);
+	// Compute samples and resize buffer
+	this->samples = clip_bytes / sizeof(float);
+	this->buffer = static_cast<float*>(std::realloc(this->buffer, this->samples * sizeof(float)));
 }
 
+// Free buffer.
 void ng::Clip::quit () {
-	if (this->data != NULL) {
-		std::free(this->data);
-		this->data = NULL;
-		this->samples = 0;
+	this->samples = 0;
+	if (this->buffer != NULL) {
+		std::free(this->buffer);
+		this->buffer = NULL;
 	}
 }
 
 void ng::Sound::init (Clip* clip, int mode) {
-	// start sound at the beginning of a clip
 	this->clip = clip;
 	this->sample = 0;
 	this->mode = mode;
 }
 
 void ng::Channel::init () {
-	this->data = NULL;
+	this->queue = NULL;
 	this->sounds = 0;
+	this->buffer = NULL;
+	this->samples = 0;
 }
 
+// Free queue and buffer.
 void ng::Channel::quit () {
-	// free sounds
-	if (this->sounds > 0) {
-		std::free(this->data);
-		this->data = NULL;
-		this->sounds = 0;
+	this->sounds = 0;
+	if (this->queue == NULL) {
+		std::free(this->queue);
+		this->queue = NULL;
+	}
+	this->samples = 0;
+	if (this->buffer == NULL) {
+		std::free(this->buffer);
+		this->buffer = NULL;
 	}
 }
 
-void ng::Channel::start_sound (Clip* clip, int mode) { // add sound
-	this->data = static_cast<Sound*>(std::realloc(this->data, (this->sounds+1) * sizeof(Sound)));
-	this->data[this->sounds].init(clip, mode);
+// Queue a sound with EnumSound mode.
+void ng::Channel::play_sound (Clip* c, int mode) {
+	this->queue = static_cast<Sound*>(std::realloc(this->queue,
+		(this->sounds + 1) * sizeof(Sound)));
+	this->queue[this->sounds].init(c, mode);
 	this->sounds++;
 }
 
-void ng::Channel::stop_sound (int sound) { // remove sound
-	if (this->sounds < 1) {
-		// no sounds to remove
-		return;
-	} else if (this->sounds == 1) {
-		// remove last sound
-		std::free(this->data);
-		this->data = NULL;
-		this->sounds = 0;
+// Remove queued sound.
+void ng::Channel::remove_sound (int sound) {
+	if (this->sounds == 0 || sound < 0 || sound >= this->sounds) {
 		return;
 	}
-	// remove sound, shift remaining sounds to fill void, and realloc
-	for (int i = sound; i < this->sounds-1; i++) {
-		this->data[i] = this->data[i+1];
-	}
-	this->data = static_cast<Sound*>(std::realloc(this->data, (this->sounds-1) * sizeof(Sound)));
-	this->sounds--;
-}
-
-void ng::Channel::stop () { // remove all sounds
-	if (this->sounds > 0) {
-		std::free(this->data);
-		this->data = NULL;
+	
+	if (this->sounds == 1) {
+		// Remove last sound and free queue.
+		std::free(this->queue);
+		this->queue = NULL;
 		this->sounds = 0;
-	}
-}
-
-void ng::Audio::init () {
-	SDL_AudioDeviceID device_id = 0;
-	SDL_AudioSpec desired;
-	desired.freq = 44100;
-	desired.format = AUDIO_U8;
-	desired.samples = 4096;
-	desired.channels = 1;
-	desired.callback = NULL;
-	SDL_AudioSpec obtained;
-	device_id = SDL_OpenAudioDevice(NULL, 0, &desired, &obtained, 0);
-	if (device_id == 0) {
-		throw std::runtime_error(SDL_GetError());
-	}
-	
-	this->spec = obtained;
-	this->device = device_id;
-	this->data = static_cast<uint8_t*>(std::malloc(4096 * sizeof(uint8_t)));
-	this->samples = 4096;
-	this->playing = false;
-}
-
-void ng::Audio::quit () {
-	// free queue and clear queued audio
-	std::free(this->data);
-	this->data = NULL;
-	this->samples = 0;
-	SDL_ClearQueuedAudio(this->device);
-	
-	// close audio device
-	SDL_CloseAudioDevice(this->device);
-	this->device = 0;
-}
-
-void ng::Audio::play () { // unpause audio device
-	SDL_PauseAudioDevice(this->device, 0);
-	this->playing = true;
-}
-
-void ng::Audio::pause () { // pause audio device
-	SDL_PauseAudioDevice(this->device, 1);
-	this->playing = false;
-}
-
-void ng::Audio::stop () { // clear queued audio
-	SDL_ClearQueuedAudio(this->device);
-}
-
-void ng::Audio::clear () { // fill queue with silence
-	if (this->playing && SDL_GetQueuedAudioSize(this->device) < (this->samples * 2)) {
-		int silence = 0;
-		for (int i = 0; i < static_cast<int>(this->samples); i++) {
-			this->data[i] = silence;
+	} else {
+		// Shift remaining sounds to fill gap and reduce queue by 1.
+		for (int i=sound; i < this->sounds - 1; i++) {
+			this->queue[i] = this->queue[i + 1];
 		}
+		this->queue = static_cast<Sound*>(std::realloc(this->queue,
+			(this->sounds - 1) * sizeof(Sound)));
+		this->sounds--;
 	}
 }
 
-int ng::Audio::mix_sample (int a, int b) const { // add decibels
-	int sum;
-	sum = (a+b)/2;
-	return sum;
+// Remove all queued sounds.
+void ng::Channel::stop () {
+	if (this->queue != NULL) {
+		std::free(this->queue);
+		this->queue = NULL;
+		this->sounds = 0;
+	}
 }
 
-void ng::Audio::mix_sound (Channel* c, int sound) { // add clip data to queue
-	Clip* clip = c->data[sound].clip;
-	int s = c->data[sound].sample;
-	switch (c->data[sound].mode) {
+// Allocate samples for buffer and fill with silence.
+void ng::Channel::clear (int samples) {
+	this->buffer = static_cast<float*>(realloc(this->buffer, samples * sizeof(float)));
+	this->samples = samples;
+	for (int i=0; i < samples; i++) {
+		this->buffer[i] = ng::dB_silence();
+	}
+}
+
+// Mix samples from sound clip into channel buffer.
+int ng::Channel::mix_sound (int sound) {
+	Clip* clip = this->queue[sound].clip;
+	int s = this->queue[sound].sample;
+	
+	switch (this->queue[sound].mode) {
 		case ng::SoundPlayOnce: {
-			for (int i=0; i < static_cast<int>(this->samples); i++) {
-				this->data[i] = this->mix_sample(this->data[i], clip->data[s]);
+			for (int i=0; i < this->samples; i++) {
+				this->buffer[i] = ng::mix_dB(this->buffer[i], clip->buffer[s]);
 				s++;
-				if (s >= static_cast<int>(clip->samples)) {
-					c->data[sound].mode = ng::SoundComplete;
-					break;
+				// Returns ng::SoundComplete if sound completes during the mix.
+				if (s == clip->samples) {
+					this->queue[sound].mode = ng::SoundComplete;
+					this->queue[sound].sample = s;
+					return ng::SoundComplete;
 				}
 			}
 			break;
+			
 		} case ng::SoundLoop: {
-			for (int i=0; i < static_cast<int>(this->samples); i++) {
-				this->data[i] = this->mix_sample(this->data[i], clip->data[s]);
+			for (int i=0; i < this->samples; i++) {
+				this->buffer[i] = ng::mix_dB(this->buffer[i], clip->buffer[s]);
 				s++;
-				if (s >= static_cast<int>(clip->samples)) {
+				if (s == clip->samples) {
 					s = 0;
 				}
 			}
 			break;
 		}
 	}
-	c->data[sound].sample = s;
+	
+	this->queue[sound].sample = s;
+	return ng::None;
 }
 
-void ng::Audio::mix_channel (Channel* c) {
-	if (this->playing && SDL_GetQueuedAudioSize(this->device) < (this->samples * 2)) {
-		for (int i = 0; i < c->sounds;) {
-			this->mix_sound(c, i);
-			if (c->data[i].mode == ng::SoundComplete) {
-				c->stop_sound(i);
-			} else {
-				i++;
-			}
+// Mix sounds into channel buffer.
+void ng::Channel::mix () {
+	for (int i=0; i < sounds;) {
+		if (this->mix_sound(i) == ng::SoundComplete) {
+			this->remove_sound(i);
+		} else {
+			i++;
 		}
 	}
 }
 
-void ng::Audio::queue () { // queue audio
-	if (this->playing && SDL_GetQueuedAudioSize(this->device) < (this->samples * 2)) {
-		if (SDL_QueueAudio(this->device, this->data, this->samples) != 0) {
+// Open audio device in a paused state and set playing to false.
+void ng::Audio::init () {
+	this->buffer = NULL;
+	this->samples = 0;
+	this->playing = false;
+	
+	SDL_AudioSpec desired, obtained;
+	desired.freq = 44100;
+	desired.format = AUDIO_F32SYS;
+	desired.samples = 4096;
+	desired.channels = 1;
+	desired.callback = NULL;
+	//int allowed_changes =
+	//	static_cast<int>(SDL_AUDIO_ALLOW_FREQUENCY_CHANGE) |
+	//	static_cast<int>(SDL_AUDIO_ALLOW_CHANNELS_CHANGE) |
+	//	static_cast<int>(SDL_AUDIO_ALLOW_SAMPLES_CHANGE);
+	this->device = SDL_OpenAudioDevice(NULL, 0, &desired, &obtained, 0);
+	if (this->device == 0) {
+		throw std::runtime_error(SDL_GetError());
+	}
+	
+	this->spec = obtained;
+}
+
+// Free buffer and close audio device.
+void ng::Audio::quit () {
+	this->playing = false;
+	this->samples = 0;
+	if (this->buffer!= NULL) {
+		std::free(this->buffer);
+		this->buffer = NULL;
+	}
+	SDL_ClearQueuedAudio(this->device);
+	SDL_CloseAudioDevice(this->device);
+	this->device = 0;
+}
+
+// Allocate at least ms of samples for buffer and fill with silence.
+void ng::Audio::clear (int ms) {
+	int queue_bytes, queue_samples;
+	// If playing is false, then does not check audio device.
+	if (this->playing) {
+		queue_bytes = static_cast<int>(SDL_GetQueuedAudioSize(this->device));
+		queue_samples = queue_bytes / sizeof(float);
+	} else {
+		queue_bytes = 0;
+		queue_samples = 0;
+	}
+	
+	int chunk_samples, samples_per_ms, samples;
+	chunk_samples = this->spec.samples * this->spec.channels;
+	samples_per_ms = this->spec.freq / 1000;
+	// If ms == 0, allocates 1 chunk of (spec.samples * spec.channels) samples.
+	if (ms > 0) {
+		samples = ms * samples_per_ms;
+	} else {
+		samples = chunk_samples;
+	}
+	// If the audio device contains more samples than ms would create, then
+	// sets samples = 0 and does nothing to buffer.
+	if (queue_samples >= samples) {
+		this->samples = 0;
+		return;
+	}
+	
+	// Samples is a multiple of (spec.samples * spec.channels) chunks, and
+	// takes into account the number of samples already in audio device.
+	samples -= queue_samples;
+	int chunks = (samples / chunk_samples) + 1;
+	samples = chunks * chunk_samples;
+	
+	this->buffer = static_cast<float*>(std::realloc(this->buffer, samples * sizeof(float)));
+	this->samples = samples;
+	for (int i=0; i < samples; i++) {
+		this->buffer[i] = ng::dB_silence();
+	}
+}
+
+// Clear channel, mix sounds, and mix channel buffer.
+void ng::Audio::mix_channel (Channel* c) {
+	// Does nothing if samples == 0.
+	if (this->samples == 0) {
+		return;
+	}
+	
+	c->clear(this->samples); // Channel samples = samples.
+	c->mix();
+	for (int i=0; i < this->samples; i++) {
+		this->buffer[i] = ng::mix_dB(this->buffer[i], c->buffer[i]);
+	}
+}
+
+// Send buffer to audio device and set playing to true.
+void ng::Audio::play () {
+	// Does nothing if samples == 0.
+	if (this->samples == 0) {
+		return;
+	}
+	
+	// Send buffer in chunks of spec.samples * spec.channels.
+	int chunk_samples, chunks, chunk_bytes;
+	chunk_samples = static_cast<int>(this->spec.samples) *
+		static_cast<int>(this->spec.channels);
+	chunks = this->samples / chunk_samples;
+	chunk_bytes = chunk_samples * sizeof(float);
+	
+	float* chunk = static_cast<float*>(std::malloc(chunk_bytes));
+	for (int i=0; i < chunks; i++) {
+		for (int f=0; f < chunk_samples; f++) {
+			chunk[f] = this->buffer[(i * chunk_samples) + f];
+		}
+		int result = SDL_QueueAudio(this->device, chunk, static_cast<uint32_t>(chunk_bytes));
+		if (result != 0) {
 			throw std::runtime_error(SDL_GetError());
 		}
+	}
+	std::free(chunk);
+	
+	// The first call to play sets playing to true and unpauses audio device.
+	if (!this->playing) {
+		SDL_PauseAudioDevice(this->device, 0); // unpause audio device
+		this->playing = true;
 	}
 }